#13 Address review: per-op timeouts, extract coalesce, location comment
- Use per-operation 30s timeout in drive handlers instead of captured
  parent context (matches telemetry handler pattern)
- Extract coalesce + mergeUpdate to writer_coalesce.go (writer.go
  down from 295 to 236 lines)
- Add safety comment on location pointer referencing event payload
